Reformatting for clarity
