+++ tests/Rewards/test_purchaseNFT.py
